+++ Web/BeatsWave.Web.ViewModels/Home/IndexBeatViewModel.cs
@@ -18,7 +18,7 @@
 
         public int BeatId { get; set; }
 
-        public decimal StandartPrice { get; set; }
+        public int StandartPrice { get; set; }
 
         public string Description { get; set; }
