+++ Web/BeatsWave.Web.ViewModels/Users/BeatToBuyViewModel.cs
@@ -20,7 +20,7 @@
 
         public string BeatUrl { get; set; }
 
-        public decimal StandartPrice { get; set; }
+        public int StandartPrice { get; set; }
 
         public string Description { get; set; }
